+++ src/pages/formPage/components/InfoForm.tsx
@@ -1,13 +1,13 @@
-import React, { useEffect, useMemo, useRef, useState } from 'react';
-import s from './InfoForm.module.scss';
-import BasicButton from '../../../components/BasicButton';
-import StatusSelect from '../../../components/StatusSelect';
-import { AcademicStatus, StepComponentProps } from '../../../types/form';
+import React, { useMemo } from "react";
+import s from "./InfoForm.module.scss";
+import BasicButton from "../../../components/BasicButton";
+import StatusSelect from "../../../components/StatusSelect";
+import { AcademicStatus, StepComponentProps } from "../../../types/form";
 import {
   ACADEMIC_STATUS_OPTIONS,
   REGION_OPTIONS,
-  SIGUNGU_MAP,
-} from '../../../data/formData';
+} from "../../../data/formData";
+import RegionSelect from "./RegionSelect";
 
 const InfoForm: React.FC<StepComponentProps> = ({
   data,
@@ -18,7 +18,7 @@ const InfoForm: React.FC<StepComponentProps> = ({
     // 재학
     updateForm({
       isEnrolled: true,
-      university: data.university ?? '',
+      university: data.university ?? "",
     });
   };
 
@@ -33,7 +33,7 @@ const InfoForm: React.FC<StepComponentProps> = ({
 
   // 필수값 충족 여부
   const ageOk = useMemo(() => {
-    const onlyNums = String(data.age ?? '').trim();
+    const onlyNums = String(data.age ?? "").trim();
     return /^[0-9]+$/.test(onlyNums) && Number(onlyNums) > 0;
   }, [data.age]);
 
@@ -72,7 +72,7 @@ const InfoForm: React.FC<StepComponentProps> = ({
                 value={data.age}
                 onChange={(e) => {
                   // 백엔드한테 넘길때 숫자로 변환 필요
-                  const onlyNums = e.target.value.replace(/[^0-9]/g, '');
+                  const onlyNums = e.target.value.replace(/[^0-9]/g, "");
                   updateForm({ age: onlyNums });
                 }}
                 aria-label="나이"
@@ -83,8 +83,8 @@ const InfoForm: React.FC<StepComponentProps> = ({
             <div className={s.col}>
               <div className={s.subLabel}>사업장 주소</div>
               <RegionSelect
-                city={data.addressCity || ''}
-                district={data.addressDistrict || ''}
+                city={data.addressCity || ""}
+                district={data.addressDistrict || ""}
                 onChange={(city, district) =>
                   updateForm({ addressCity: city, addressDistrict: district })
                 }
@@ -132,7 +132,7 @@ const InfoForm: React.FC<StepComponentProps> = ({
                 className={s.unv}
                 type="text"
                 placeholder="학교입력"
-                value={data.university ?? ''}
+                value={data.university ?? ""}
                 onChange={(e) => updateForm({ university: e.target.value })}
                 aria-label="대학교"
               />
@@ -161,145 +161,5 @@ const InfoForm: React.FC<StepComponentProps> = ({
     </>
   );
 };
-
-/* ===== 지역 선택 드롭다운 (단일 선택) ===== */
-function RegionSelect({
-  city,
-  district,
-  onChange,
-  placeholder = '지역 선택하기',
-  options,
-}: {
-  city: string;
-  district: string;
-  onChange: (city: string, district: string) => void;
-  placeholder?: string;
-  options: string[];
-}) {
-  const [open, setOpen] = useState(false);
-  const ref = useRef<HTMLDivElement>(null);
-
-  // 단일 선택 상태
-  const [selectedSido, setSelectedSido] = useState<string | null>(null);
-  const [selectedSigungu, setSelectedSigungu] = useState<string | null>(null);
-  const [activeSido, setActiveSido] = useState<string | null>(null);
-
-  useEffect(() => {
-    if (city) {
-      setSelectedSido(city);
-      setActiveSido(city);
-    } else {
-      setSelectedSido(null);
-      setActiveSido(null);
-    }
-    setSelectedSigungu(district || null);
-  }, [city, district]);
-
-  useEffect(() => {
-    const onDoc = (e: MouseEvent) => {
-      if (ref.current && !ref.current.contains(e.target as Node))
-        setOpen(false);
-    };
-    document.addEventListener('mousedown', onDoc);
-    return () => document.removeEventListener('mousedown', onDoc);
-  }, []);
-
-  const label = useMemo(() => {
-    if (selectedSido)
-      return selectedSigungu
-        ? `${selectedSido} / ${selectedSigungu}`
-        : selectedSido;
-    return (city && district ? `${city} / ${district}` : '') || placeholder;
-  }, [selectedSido, selectedSigungu, city, district, placeholder]);
-
-  const chooseSido = (sido: string) => {
-    setSelectedSido(sido);
-    setActiveSido(sido);
-    setSelectedSigungu(null);
-  };
-
-  const chooseSigungu = (sgg: string) => {
-    if (!selectedSido) return;
-    setSelectedSigungu(sgg);
-    onChange(selectedSido, sgg);
-    setOpen(false);
-  };
-
-  const sigunguList = useMemo(
-    () => (activeSido ? SIGUNGU_MAP[activeSido] || [] : []),
-    [activeSido]
-  );
-
-  return (
-    <div className={s.regionWrap} ref={ref}>
-      <button
-        type="button"
-        className={`${s.regionTrigger} ${
-          city && district ? s.regionTriggerActive : ''
-        }`}
-        onClick={() => setOpen((o) => !o)}
-        aria-haspopup="listbox"
-        aria-expanded={open}
-      >
-        <span className={s.regionLabel}>{label}</span>
-      </button>
-
-      {open && (
-        <>
-          {/* 시/도 */}
-          <div
-            className={s.regionMenu}
-            role="listbox"
-            onMouseDown={(e) => e.preventDefault()}
-          >
-            <div className={s.selectext}>시/도</div>
-            {options.map((sido) => {
-              const checked = selectedSido === sido;
-              return (
-                <button
-                  key={sido}
-                  type="button"
-                  className={s.regionOption}
-                  onClick={() => chooseSido(sido)}
-                  aria-selected={checked}
-                >
-                  <span className={`${s.box} ${checked ? s.boxOn : ''}`} />
-                  <span className={s.regionText}>{sido}</span>
-                </button>
-              );
-            })}
-          </div>
-
-          {/* 시/군/구 */}
-          {activeSido && (
-            <div
-              className={s.regionMenu}
-              role="listbox"
-              onMouseDown={(e) => e.preventDefault()}
-              style={{ left: 'calc(100% + 10vw)' }}
-            >
-              <div className={s.selectext}>시/군/구</div>
-              {sigunguList.map((sgg) => {
-                const checked = selectedSigungu === sgg;
-                return (
-                  <button
-                    key={sgg}
-                    type="button"
-                    className={s.regionOption}
-                    onClick={() => chooseSigungu(sgg)}
-                    aria-selected={checked}
-                  >
-                    <span className={`${s.box} ${checked ? s.boxOn : ''}`} />
-                    <span className={s.regionText}>{sgg}</span>
-                  </button>
-                );
-              })}
-            </div>
-          )}
-        </>
-      )}
-    </div>
-  );
-}
 
 export default InfoForm;
